Fix D9025 and C26451 warnings

+++ VTIL-Common/io/formatting.hpp
@@ -117,8 +117,8 @@ namespace vtil::format
 					return fix_type_name( in.substr( strlen( str ) ) );
 				for ( int i = 0; i < in.size(); i++ )
 				{
-					if ( in[ i ] == '<' && in.substr( i + 1 ).starts_with( str ) )
-						in = in.substr( 0, i + 1 ) + in.substr( i + 1 + strlen( str ) );
+					if ( in[ i ] == '<' && in.substr( size_t( i ) + 1 ).starts_with( str ) )
+						in = in.substr( 0, size_t( i ) + 1 ) + in.substr( size_t( i ) + 1 + strlen( str ) );
 				}
 			}
 			return in;
